// src/core/types/doctorProfile.ts
export interface EducationEntry {
  id: number;
  degree: string;
  university: string;
  from: string;
  to: string;
}

export interface AwardEntry {
  id: number;
  name: string;
  description: string;
  year: string;
}

export interface CertificationEntry {
  id: number;
  name: string;
  description: string;
  year: string;
}